requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-resized==0.3.11
django-rest-swagger==2.2.0
django-stdimage==5.3.0
djangorestframework==3.12.4
django-choice=1.7.2
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.2
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.7
